fix(gb28181): stream monitor FPS 精度与失败日志
- ZLM MediaTrack.FPS 与 Snapshot.Track.FPS 改为 float64, 保留小数以匹配 ZLM 返回
- read 失败时补 warn 日志, 记录 streamId/nodeId/host 便于排查节点不可用

+++ server/app/gb28181/zlm/client.go
@@ -151,7 +151,7 @@ type MediaTrack struct {
 	SampleBit     int      `json:"sample_bit"`
 	Width         int      `json:"width"`
 	Height        int      `json:"height"`
-	FPS           int      `json:"fps"`
+	FPS           float64  `json:"fps"`
 	KeyFrames     int64    `json:"key_frames"`
 	GOPSize       int      `json:"gop_size"`
 	GOPIntervalMS int      `json:"gop_interval_ms"`
